third_party/WebKit/LayoutTests/fast/css-intrinsic-dimensions/change-inside-percent-padding.html

<!DOCTYPE html>
<p>There should be a blue square below.</p>
<div id="container" style="float:left; height:100px; background:blue;" data-expected-width="100">
    <div style="padding-right:10%;">
        <div id="boo" style="display:none; width:100px; height:1px;"></div>
        <div style="width:10px; height:1px;"></div>
    </div>
</div>
<div style="clear:both;"></div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<script>
    document.body.offsetTop;
    document.getElementById("boo").style.display = "block";
    checkLayout("#container");
</script>
